🐎 refactor: make the FOUT actually happen
The loadCSS call to custom-fonts.css is well-intentioned, but it just
delays the FOUT until later, causing the system fonts to blink out when
the stylesheet asynchronously loads; this is technically even worse than
the default behavior of having a raw @font-face in the sheet.

With a little help from zenfonts.js, the FOUT is here for real.